RDKCOM-344: devices:properties: add default value for PROCESS_RESTART_LOG
This value is used in ccsp systemd scripts and it's not defined anywhere
for qemu or rpi3 targets (note that rpi3 currently uses this tree for
it's devices property file).
We can define a default not file to /dev/null so we don't get spammed
with errors when a service is restarting.
Signed-off-by: Matthew McClintock <msm-oss@mcclintock.net>

Reason for change: devices:properties: add default value for PROCESS_RESTART_LOG
Test Procedure: None.
Risks: Low.

Change-Id: I196d02b153172487e78e88a030bc9c0f794fee96